Refactoring and tweaking widget
